requirements.txt
PyGithub
slackclient